pyproject.toml
[project]
name = "beschess"
version = "0.1.0"
description = "Add your description here"
authors = [ ]
requires-python = ">=3.13"
dependencies = [
    "backports-zstd>=1.0.0",
    "h5py>=3.15.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "python-chess>=1.999",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "stockfish>=3.28.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "ujson>=5.11.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beschess"]

[tool.ruff]
line-length = 88
exclude=[
    ".eggs",
    ".git",
    ".mypy_cache",
    ".vscode",
    "__pycache__",
    "build",
    "dist",
    "sklearn/externals",
    "doc/_build",
    "doc/auto_examples",
    "asv_benchmarks/env",
    "asv_benchmarks/html",
    "asv_benchmarks/results",
    "asv_benchmarks/benchmarks/cache",
]
